# Conflicts:
#	models/src/main/java/ar/edu/itba/paw/models/JobPost.java

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/jpa/JobCardDaoJpa.java
@@ -96,14 +96,14 @@ public class JobCardDaoJpa implements JobCardDao {
 
     @Override
     public int findAllMaxPage() {
-        BigInteger size = em.createQuery("SELECT COUNT(*) FROM JobPost jpost WHERE jpost.isActive = TRUE", BigInteger.class)
+        Long size = em.createQuery("SELECT COUNT(*) FROM JobPost jpost WHERE jpost.isActive = TRUE", Long.class)
                 .getSingleResult();
         return (int) Math.ceil((double) size.intValue() / HirenetUtils.PAGE_SIZE);
     }
 
     @Override
     public int findMaxPageByUserId(long id) {
-        BigInteger size = em.createQuery("SELECT COUNT(*) FROM JobPost jpost WHERE jpost.user.id = :id AND jpost.isActive = TRUE", BigInteger.class)
+        Long size = em.createQuery("SELECT COUNT(*) FROM JobPost jpost WHERE jpost.user.id = :id AND jpost.isActive = TRUE", Long.class)
                 .setParameter("id", id).getSingleResult();
         return (int) Math.ceil((double) size.intValue() / HirenetUtils.PAGE_SIZE);
     }
